# libs/domains/agenstra/backend/feature-agent-controller/src/lib/controllers/knowledge-tree.controller.ts
import { RequireScopes, type RequestWithUser } from '@forepath/identity/backend';
import {
  Body,
  Controller,
  DefaultValuePipe,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  ParseIntPipe,
  ParseUUIDPipe,
  Patch,
  Post,
  Query,
  Req,
} from '@nestjs/common';

import {
  CreateKnowledgeNodeDto,
  CreateKnowledgeRelationDto,
  KnowledgePromptContextResponseDto,
  ReorderKnowledgeNodeDto,
  UpdateKnowledgeNodeDto,
} from '../dto/knowledge';
import { KnowledgeRelationSourceType } from '../entities/knowledge-node.enums';
import { KnowledgeTreeService } from '../services/knowledge-tree.service';

@Controller('knowledge')
export class KnowledgeTreeController {
  constructor(private readonly knowledgeTreeService: KnowledgeTreeService) {}

  @RequireScopes('knowledge:read')
  @Get()
  async listByClient(
    @Query('clientId', ParseUUIDPipe) clientId: string,
    @Query('search') search?: string,
    @Req() req?: RequestWithUser,
  ) {
    return await this.knowledgeTreeService.listNodes(clientId, req, search);
  }

  @RequireScopes('knowledge:read')
  @Get('tree')
  async getTree(
    @Query('clientId', ParseUUIDPipe) clientId: string,
    @Query('search') search?: string,
    @Req() req?: RequestWithUser,
  ) {
    return await this.knowledgeTreeService.getTree(clientId, req, search);
  }

  @RequireScopes('knowledge:read')
  @Get('by-sha')
  async getBySha(
    @Query('clientId', ParseUUIDPipe) clientId: string,
    @Query('sha') sha: string,
    @Req() req?: RequestWithUser,
  ) {
    return await this.knowledgeTreeService.findNodeBySha(clientId, sha, req);
  }

  @RequireScopes('knowledge:write')
  @Post()
  @HttpCode(HttpStatus.CREATED)
  async createNode(@Body() dto: CreateKnowledgeNodeDto, @Req() req?: RequestWithUser) {
    return await this.knowledgeTreeService.createNode(dto, req);
  }

  @RequireScopes('knowledge:write')
  @Post('relations')
  @HttpCode(HttpStatus.CREATED)
  async createRelation(@Body() dto: CreateKnowledgeRelationDto, @Req() req?: RequestWithUser) {
    return await this.knowledgeTreeService.createRelation(dto, req);
  }

  @RequireScopes('knowledge:read')
  @Get('relations')
  async listRelations(
    @Query('clientId', ParseUUIDPipe) clientId: string,
    @Query('sourceType') sourceType: KnowledgeRelationSourceType,
    @Query('sourceId', ParseUUIDPipe) sourceId: string,
    @Req() req?: RequestWithUser,
  ) {
    return await this.knowledgeTreeService.listRelations(clientId, sourceType, sourceId, req);
  }

  @RequireScopes('knowledge:write')
  @Delete('relations/:id')
  @HttpCode(HttpStatus.NO_CONTENT)
  async deleteRelation(@Param('id', ParseUUIDPipe) id: string, @Req() req?: RequestWithUser) {
    await this.knowledgeTreeService.deleteRelation(id, req);
  }

  @RequireScopes('knowledge:read')
  @Get('relations/prompt-context')
  async getPromptContexts(
    @Query('clientId', ParseUUIDPipe) clientId: string,
    @Query('sourceType') sourceType: KnowledgeRelationSourceType,
    @Query('sourceId', ParseUUIDPipe) sourceId: string,
    @Req() req?: RequestWithUser,
  ): Promise<KnowledgePromptContextResponseDto> {
    return await this.knowledgeTreeService.collectPromptContextsForSource(clientId, sourceType, sourceId, req);
  }

  @RequireScopes('knowledge:read')
  @Get(':id')
  async getNode(@Param('id', ParseUUIDPipe) id: string, @Req() req?: RequestWithUser) {
    return await this.knowledgeTreeService.getNode(id, req);
  }

  @RequireScopes('knowledge:read')
  @Get(':id/activity')
  async listActivity(
    @Param('id', ParseUUIDPipe) id: string,
    @Query('limit', new DefaultValuePipe(50), ParseIntPipe) limit: number,
    @Query('offset', new DefaultValuePipe(0), ParseIntPipe) offset: number,
    @Req() req?: RequestWithUser,
  ) {
    return await this.knowledgeTreeService.listPageActivity(id, limit, offset, req);
  }

  @RequireScopes('knowledge:write')
  @Patch(':id')
  async updateNode(
    @Param('id', ParseUUIDPipe) id: string,
    @Body() dto: UpdateKnowledgeNodeDto,
    @Req() req?: RequestWithUser,
  ) {
    return await this.knowledgeTreeService.updateNode(id, dto, req);
  }

  @RequireScopes('knowledge:write')
  @Post(':id/reorder')
  async reorderNode(
    @Param('id', ParseUUIDPipe) id: string,
    @Body() dto: ReorderKnowledgeNodeDto,
    @Req() req?: RequestWithUser,
  ) {
    return await this.knowledgeTreeService.reorderNode(id, dto, req);
  }

  @RequireScopes('knowledge:write')
  @Post(':id/duplicate')
  async duplicateNode(@Param('id', ParseUUIDPipe) id: string, @Req() req?: RequestWithUser) {
    return await this.knowledgeTreeService.duplicateNode(id, req);
  }

  @RequireScopes('knowledge:write')
  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  async deleteNode(
    @Param('id', ParseUUIDPipe) id: string,
    @Query('releaseExternalSyncMarker') releaseExternalSyncMarkerRaw?: string,
    @Req() req?: RequestWithUser,
  ) {
    const releaseExternalSyncMarker = releaseExternalSyncMarkerRaw === 'true' || releaseExternalSyncMarkerRaw === '1';

    await this.knowledgeTreeService.deleteNode(id, req, releaseExternalSyncMarker);
  }
}